Adding confirmation page for fact check
