+++ app/views/fact_check_complete.scala.html
@@ -1,0 +1,33 @@
+@*
+ * Copyright 2020 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import config.AppConfig
+
+@import uk.gov.hmrc.govukfrontend.views.html.components._
+
+@this(layout: main_layout)
+
+@()(implicit request: Request[_], messages: Messages)
+
+@layout(messages("factCheckComplete.title")) {
+
+    <h1 class="govuk-heading-xl">@messages("factCheckComplete.heading")</h1>
+    <p class="govuk-body">@messages("factCheckComplete.info")</p>
+    <a href="@{routes.AdminController.approvalSummaries.toString}" aria-describedby="close" class="govuk-button" data-module="govuk-button">@messages("factCheckComplete.close")</a>
+}
+@{
+ //$COVERAGE-OFF$
+}
