+++ app/views/fact_check_content_review.scala.html
@@ -65,7 +65,7 @@
 
       <ul class="app-task-list__items">
         <li class="app-task-list__item">
-          <span class="app-task-list__task-name"><a href="@{routes.FactCheckController.onConfirm(review.ocelotId).toString}" aria-describedby="send-confirmation">@messages("factCheck.sendConfirmation")</a></span>
+          <span class="app-task-list__task-name"><a href="@{routes.FactCheckConfirmController.onConfirm(review.ocelotId).toString}" aria-describedby="send-confirmation">@messages("factCheck.sendConfirmation")</a></span>
           <span class="hmrc-status-tag" id="send-confirmation">@messages("factCheck.sendConfirmationStatus")</span>
         </li>
       </ul>
