+++ conf/app.routes
@@ -1,14 +1,14 @@
 # microservice specific routes
 GET        /assets/*file                              controllers.Assets.versioned(path="/public", file: Asset)
-        
+
 GET        /process/approval                          controllers.AdminController.approvalSummaries
 GET        /2i-review/:id                             controllers.TwoEyeReviewController.approval(id)
-        
+
 GET        /2i-page-review/:id/*page                  controllers.TwoEyePageReviewController.onPageLoad(id: String, page: String)
 GET        /2i-page-review/:id/                       controllers.TwoEyePageReviewController.onPageLoad(id: String, page: String ?= "")
 POST       /2i-page-review/:id/*page                  controllers.TwoEyePageReviewController.onSubmit(id: String, page: String )
 POST       /2i-page-review/:id/                       controllers.TwoEyePageReviewController.onSubmit(id: String, page: String ?= "")
-        
+
 GET        /2i-result/:id                             controllers.TwoEyeReviewResultController.onPageLoad(id)
 POST       /2i-result/:id                             controllers.TwoEyeReviewResultController.onSubmit(id)
 
@@ -18,22 +18,22 @@ POST       /fact-check-page-review/:id/*page          controllers.FactCheckPageR
 POST       /fact-check-page-review/:id/               controllers.FactCheckPageReviewController.onSubmit(id: String, page: String ?= "")
 
 GET        /fact-check/:id                            controllers.FactCheckController.approval(id)
-GET        /fact-check-result/:id                     controllers.FactCheckController.onConfirm(id)
-        
+GET        /fact-check-result/:id                     controllers.FactCheckConfirmController.onConfirm(id)
+
 GET        /accessibility                             controllers.AccessibilityStatementController.getPage
-        
+
 GET        /hello-world                               controllers.HelloWorldController.helloWorld
 GET        /bye-world                                 controllers.HelloWorldController.byeWorld
-        
+
 POST       /process/scratch                           controllers.apis.ScratchController.submitScratchProcess
 OPTIONS    /process/scratch                           controllers.apis.ScratchController.scratchProcessOptions
-        
+
 POST       /process/approval                          controllers.apis.ApprovalController.submitFor2iReview
 OPTIONS    /process/approval                          controllers.apis.ApprovalController.options
-        
+
 POST       /process/approval/2i-review                controllers.apis.ApprovalController.submitFor2iReview
 OPTIONS    /process/approval/2i-review                controllers.apis.ApprovalController.options
-        
+
 POST       /process/approval/fact-check               controllers.apis.ApprovalController.submitForFactCheck
 OPTIONS    /process/approval/fact-check               controllers.apis.ApprovalController.options
