+++ test/controllers/FactCheckControllerSpec.scala
@@ -141,118 +141,4 @@ class FactCheckControllerSpec extends ControllerBaseSpec with GuiceOneAppPerSuit
 
   }
 
-  "The fact check controller when confirming the review" should {
-
-    "Return SEE_OTHER for a successful post of the review completion for a process" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Right(())))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      status(result) shouldBe Status.SEE_OTHER
-    }
-
-    "Return an Html document displaying the details of the review result" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Right(())))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      contentType(result) shouldBe None
-    }
-
-    "Return the Http status Not found when the process review does not exist" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(NotFoundError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      status(result) shouldBe Status.NOT_FOUND
-    }
-
-    "Return an Html error page when the process review does not exist" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(NotFoundError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      contentType(result) shouldBe Some(MimeTypes.HTML)
-    }
-
-    "Return the Http status internal server error when the process review data returned to the application is malformed" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(MalformedResponseError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      status(result) shouldBe Status.INTERNAL_SERVER_ERROR
-    }
-
-    "Return an Html error page when the process review data returned to the application is malformed" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(MalformedResponseError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      contentType(result) shouldBe Some(MimeTypes.HTML)
-    }
-
-    "Return the Http status not found when the process review data requested is out of date in some way" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(StaleDataError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      status(result) shouldBe Status.NOT_FOUND
-    }
-
-    "Return an Html error page when the process review data requested is out of date in some way" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(StaleDataError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      contentType(result) shouldBe Some(MimeTypes.HTML)
-    }
-
-    "Return the Http status internal server error when an unexpected error occurs" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(InternalServerError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      status(result) shouldBe Status.INTERNAL_SERVER_ERROR
-    }
-
-    "Return an Html error page when an unexpected error occurs" in new Test {
-
-      MockReviewService
-        .approvalFactCheckComplete(id, credential, name, ApprovalStatus.WithDesignerForUpdate)
-        .returns(Future.successful(Left(InternalServerError)))
-
-      val result: Future[Result] = reviewController.onConfirm(id)(fakeRequest)
-
-      contentType(result) shouldBe Some(MimeTypes.HTML)
-    }
-
-  }
-
 }
